first, fully working commit
This is *not* the first commit; it's only the first commit of the
re-incarnation known as ePread. The first actual commit was almost
3 years ago, in 2018, to make an application like Rambox and others,
then it's grown in features and "modes". One of these modes was the
EPUB mode. Other modes include Tabbed Window mode (similar to Rambox)
and Single Window mode (each "app" in its window, optionally with a
system tray icon to close to), and Print mode (to convert html pages,
whether online or local, to PDF).

That project was so complex (but not large) that I didn't like to
upload, and probably never will. Although I still use it almost
daily for some of the web apps I use.
Extracting useful "sub-projects" from it is the way to go I think.